requirements/requirements.txt
# Minimal requirements for Multi-Stage Fusion Docker submission
# Optimized for size and performance

# Core dependencies
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Tracking and filtering
filterpy==1.4.5
scikit-learn==1.3.2

# Deep learning utilities
Pillow==10.1.0
tqdm==4.66.1
pyyaml==6.0.1

# YOLO-Pose for keypoint detection
ultralytics==8.3.0
